{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug }}/api/mixins.py
from typing import TYPE_CHECKING, Sequence, Type

from rest_framework.authentication import BaseAuthentication
from rest_framework.permissions import BasePermission, IsAuthenticated
from rest_framework_jwt.authentication import JSONWebTokenAuthentication

if TYPE_CHECKING:
    # This is going to be resolved in the stub library
    # https://github.com/typeddjango/djangorestframework-stubs/
    from rest_framework.permissions import _PermissionClass

    PermissionClassesType = Sequence[_PermissionClass]
else:
    PermissionClassesType = Sequence[Type[BasePermission]]


class ApiAuthMixin:
    authentication_classes: Sequence[Type[BaseAuthentication]] = [
        JSONWebTokenAuthentication,
        # ... other auth methods
    ]
    permission_classes: PermissionClassesType = (IsAuthenticated,)
