{{ cookiecutter.project_slug }}/requirements/base.txt
Django==4.1.7
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0

celery==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.4.0

whitenoise==6.3.0

django-filter==23.2
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1

drf-jwt==1.19.2

boto3==1.26.0
attrs==22.1.0

gunicorn==20.1.0
sentry-sdk==1.14.0

requests==2.30.0
